Complete retail parity fixes and presentation lifecycle handling
Reconcile order-pump reentrancy, weapon initialization and execution, path
admission, ground steering, visibility clipping, mission arguments, damage
overrides and duplicate archive lookup with the retail contracts.

Preserve camera and selection state; correct caption overflow and audio sample
identity and session deadlines. Bound factory diagnostics, prewarm reachable
feature art, retire renderer sources across battle and executor changes, and
support native window controls.

Add paced battle benchmarks with state, memory and phase diagnostics and offline
analysis. Refine zoom gesture timing and momentum handling and interceptor point
quantization.

+++ internal/combat/weapon_adapter.go
@@ -97,8 +97,8 @@ func FireWeaponPoint(u *units.Unit, idx int, x, z numeric.Fixed, _ uint32) bool 
 	if s == nil || s.Weapon == nil {
 		return false
 	}
-	wx := int32(x.Raw() >> 16)
-	wz := int32(z.Raw() >> 16)
+	wx := int16(x.Raw() >> 16)
+	wz := int16(z.Raw() >> 16)
 	if wz == -32768 {
 		wz = -32767
 	}
